+++ src/main/java/com/reservation/application/servlets/CourseServlet.java
@@ -11,6 +11,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
@@ -36,10 +37,21 @@ public class CourseServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("application/json");
         PrintWriter out = response.getWriter();
-        List<Course> courses = DAO.getCourses();
-        Gson gson = new Gson();
-        String toJson = gson.toJson(courses);
-        out.println(toJson);
+
+        HttpSession session = request.getSession();
+
+        String role = (String) session.getAttribute("role");
+
+        if(role != null && role.equals("admin")){
+            List<Course> courses = DAO.getCourses();
+            Gson gson = new Gson();
+            String toJson = gson.toJson(courses);
+            out.println(toJson);
+        }
+        else{
+            out.println("Non puoi compiere questa azione");
+        }
+
         out.flush();
         out.close();
     }
@@ -48,16 +60,27 @@ public class CourseServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/plain");
         PrintWriter out = response.getWriter();
-        try {
-            String title = request.getParameter("title");
-            DAO.insertCourses(title);
-            out.println("Inserimento effettuato");
-        } catch (NumberFormatException e) {
-            out.println("Inserire un numero valido");
-        } catch (SQLException e) {
-            e.printStackTrace();
-            out.println(e.getMessage());
+
+        HttpSession session = request.getSession();
+
+        String role = (String) session.getAttribute("role");
+
+        if(role != null && role.equals("admin")){
+            try {
+                String title = request.getParameter("title");
+                DAO.insertCourses(title);
+                out.println("Inserimento effettuato");
+            } catch (NumberFormatException e) {
+                out.println("Inserire un numero valido");
+            } catch (SQLException e) {
+                e.printStackTrace();
+                out.println(e.getMessage());
+            }
         }
+        else{
+            out.println("Non puoi compiere questa azione");
+        }
+
         out.flush();
         out.close();
     }
@@ -66,16 +89,27 @@ public class CourseServlet extends HttpServlet {
     protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/plain");
         PrintWriter out = response.getWriter();
-        try {
-            int idCourse = Integer.parseInt(request.getParameter("idCourse"));
-            DAO.removeCourses(idCourse);
-            out.println("Corso rimosso");
-        } catch (NumberFormatException e) {
-            out.println("Inserire un numero valido");
-        } catch (SQLException e) {
-            e.printStackTrace();
-            out.println(e.getMessage());
+
+        HttpSession session = request.getSession();
+
+        String role = (String) session.getAttribute("role");
+
+        if(role != null && role.equals("admin")){
+            try {
+                int idCourse = Integer.parseInt(request.getParameter("idCourse"));
+                DAO.removeCourses(idCourse);
+                out.println("Corso rimosso");
+            } catch (NumberFormatException e) {
+                out.println("Inserire un numero valido");
+            } catch (SQLException e) {
+                e.printStackTrace();
+                out.println(e.getMessage());
+            }
         }
+        else{
+            out.println("Non puoi compiere questa azione");
+        }
+
         out.flush();
         out.close();
     }
